[IMP] l10n_my{_ubl_pint}: migrate SST and TTX fields to base module
`sst_registration_number` and `ttx_registration_number` fields were introduced in `l10n_my_ubl_pint` during the initial e-invoicing integration.

However, Sales and Service Tax (SST) and Tourism Tax (TTX) are fundamental Malaysian accounting requirements that apply to companies regardless of their participation in the LHDN e-Invoicing mandate.

This commit decouples these fields from the UBL PINT module by moving them to the base `l10n_my` module. This ensures that SST and TTX values are visible and editable within the standard Customer Invoice and Vendor Bill views, and print on standard PDF reports, without requiring the e-Invoicing module.

task-[6046974](https://www.odoo.com/odoo/all-tasks/6046974)
