+++ addons/l10n_my/models/__init__.py
@@ -1,4 +1,5 @@
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 from . import product_template
+from . import res_company
 from . import res_partner
 from . import template_my
