+++ addons/l10n_my/views/report_invoice.xml
@@ -11,13 +11,13 @@
     </template>
 
     <!-- Add the SST and TTx to the company -->
-    <template id="l10n_my_ubl_pint_external_layout_standard" inherit_id="web.company_address_list">
+    <template id="l10n_my_external_layout_standard" inherit_id="web.company_address_list">
         <xpath expr="//ul[@name='company_address_list']" position="inside">
-            <t t-if="not hide_sst_ttx_registration_number" t-call="l10n_my_ubl_pint.sst_ttx_registration_number_external_layout"/>
+            <t t-if="not hide_sst_ttx_registration_number" t-call="l10n_my.sst_ttx_registration_number_external_layout"/>
         </xpath>
     </template>
 
-    <template id="l10n_my_ubl_pint_external_layout_folder" inherit_id="web.external_layout_folder">
+    <template id="l10n_my_external_layout_folder" inherit_id="web.external_layout_folder">
         <xpath expr="//t[@t-call='web.company_address_list']" position="before">
             <t t-set="hide_sst_ttx_registration_number" t-value="True"/>
         </xpath>
