+++ addons/l10n_my/views/res_company_view.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <odoo>
-    <record id="view_company_form_inherit_l10n_my_ubl_pint" model="ir.ui.view">
-        <field name="name">res.company.form.inherit.l10n_my_ubl_pint</field>
+    <record id="view_company_form_inherit_l10n_my" model="ir.ui.view">
+        <field name="name">res.company.form.inherit.l10n_my</field>
         <field name="model">res.company</field>
         <field name="inherit_id" ref="base.view_company_form"/>
         <field name="arch" type="xml">
