+++ addons/l10n_my/views/res_partner_view.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <odoo>
-    <record id="view_partner_form_inherit_l10n_my_ubl_pint" model="ir.ui.view">
-        <field name="name">res.partner.form.inherit.l10n_my_ubl_pint</field>
+    <record id="view_partner_form_inherit_l10n_my" model="ir.ui.view">
+        <field name="name">res.partner.form.inherit.l10n_my</field>
         <field name="model">res.partner</field>
         <field name="inherit_id" ref="account.view_partner_property_form"/>
         <field name="arch" type="xml">
