+++ addons/l10n_my_ubl_pint/__manifest__.py
@@ -8,11 +8,6 @@
     The UBL PINT e-invoicing format for Malaysia is based on the Peppol International (PINT) model for Billing.
     """,
     'depends': ['account_edi_ubl_cii'],
-    'data': [
-        'views/report_invoice.xml',
-        'views/res_company_view.xml',
-        'views/res_partner_view.xml',
-    ],
     'author': 'Odoo S.A.',
     'uninstall_hook': 'uninstall_hook',
     'license': 'LGPL-3'
