+++ addons/l10n_my_ubl_pint/models/__init__.py
@@ -1,4 +1,3 @@
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 from . import account_edi_xml_pint_my
-from . import res_company
 from . import res_partner
