+++ addons/l10n_my_ubl_pint/models/res_partner.py
@@ -1,19 +1,11 @@
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
-from odoo import models, fields, api
+from odoo import models, fields
 
 
 class ResPartner(models.Model):
     _inherit = 'res.partner'
 
     invoice_edi_format = fields.Selection(selection_add=[('pint_my', "Malaysia (Peppol PINT MY)")])
-    sst_registration_number = fields.Char(
-        string="SST",
-        help="Malaysian Sales and Service Tax Number",
-    )
-    ttx_registration_number = fields.Char(
-        string="TTx",
-        help="Malaysian Tourism Tax Number",
-    )
 
     def _get_edi_builder(self, invoice_edi_format):
         # EXTENDS 'account_edi_ubl_cii'
@@ -26,7 +18,3 @@ class ResPartner(models.Model):
         formats_info = super()._get_ubl_cii_formats_info()
         formats_info['pint_my'] = {'countries': ['MY'], 'on_peppol': True}
         return formats_info
-
-    @api.model
-    def _commercial_fields(self):
-        return super()._commercial_fields() + ['sst_registration_number', 'ttx_registration_number']
